Crate typed_tableau

A typesafe table is a table where you decide ahead of time the types for each column and the type system ensures that when you create a row, the types of the cells line up properly.

You can always switch to an untyped Tableau table (and use its many functions) by calling .to_untyped().

Example

use typed_tableau::*;
// Declare our table with headers
let mut t =
    table(column::<&str>("Name").column::<usize>("Age").column::<bool>("Married"));

// Add rows
t.add_row(cell("Joe")           .cell(10)              .cell(false));
t.add_row(cell("Mary")          .cell(23)              .cell(true ));
t.add_row(cell("John")          .cell(53)              .cell(false));
t.add_row(cell("Rob")           .cell(41)              .cell(true ));
// This will fail at compile time because we're trying to stuff the wrong type (f32) into the age column (usize)
// t.add_row(cell("Rob")        .cell(41f32)           .cell(true));
// Get back the untyped Tableau table
let untyped_t = t.into_untyped();

untyped_t.display()

Structs

Cell

Column
A typed column

Table
A table with typed columns

Traits

AppendToCells
Sugar for appending anything to an HList of Cells

AppendToColumns
Sugar for appending anything to an HList of Columns

AppendWith
Sugar for appending something that is already contained onto an HList

Functions

cell
Creates a new cell

column
Builds a column with the given type and name

table
Creates a table with the given column columns.